modified arm-target just to make it compile (establish compliance with S2E-modified interfaces). Real patching still required.

+++ android-emulator/target-arm/translate.c
@@ -9186,7 +9186,7 @@ void cpu_dump_state(CPUState *env, FILE *f,
 }
 
 void gen_pc_load(CPUState *env, TranslationBlock *tb,
-                unsigned long searched_pc, int pc_pos, void *puc)
+                 uintptr_t searched_pc, int pc_pos, void *puc)
 {
     env->regs[15] = gen_opc_pc[pc_pos];
 }
